requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
blinker=1.6.2=py312haa95532_0
brotli-python=1.0.9=py312hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.9.24=haa95532_0
cairo=1.16.0=haedb8bc_5
certifi=2024.8.30=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py312haa95532_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.3.0=pypi_0
cpuonly=2.0=0
cycler=0.12.1=pypi_0
eigen=3.4.0=h59b6b97_0
expat=2.6.3=h5da7b33_0
filelock=3.13.1=py312haa95532_0
flask=3.0.3=py312haa95532_0
fontconfig=2.14.1=hb33846d_3
fonttools=4.54.1=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2024.9.0=pyhff2d567_0
glib=2.78.4=hd77b12b_0
glib-tools=2.78.4=hd77b12b_0
graphite2=1.3.14=hd77b12b_1
gst-plugins-base=1.22.3=hfc3ed34_1
gstreamer=1.22.3=h1779075_1
harfbuzz=4.3.0=hb646838_2
hdf5=1.12.1=h51c971a_3
huggingface_hub=0.25.2=pyh0610db2_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.7=py312haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
itsdangerous=2.2.0=py312haa95532_0
jinja2=3.1.4=py312haa95532_0
jpeg=9e=h827c3e9_3
kiwisolver=1.4.7=pypi_0
krb5=1.20.1=h5b6d351_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libabseil=20240116.2=cxx17_h5da7b33_0
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libglib=2.78.4=ha17d25a_0
libiconv=1.16=h2bbff1b_3
libjpeg-turbo=2.0.0=h196d8e1_0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libprotobuf=4.25.3=hf2fb9eb_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libvorbis=1.3.7=he774522_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.13.1=h24da03e_2
lightning-utilities=0.11.7=pyhd8ed1ab_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py312h2bbff1b_0
matplotlib=3.9.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.10=py312h827c3e9_0
mkl_random=1.2.7=py312h0158946_0
mpmath=1.3.0=py312haa95532_0
networkx=3.2.1=py312haa95532_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
opencv=4.10.0=py312h4ccfa20_0
openjpeg=2.5.2=hae555c5_0
openssl=3.3.2=h2466b09_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.3=pypi_0
pcre2=10.42=h0ff8eda_1
pillow=10.4.0=py312h827c3e9_0
pip=24.2=py312haa95532_0
pixman=0.40.0=h2bbff1b_1
psutil=6.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyparsing=3.1.4=pypi_0
pysocks=1.7.1=py312haa95532_0
python=3.12.7=h14ffc60_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.4.1=py3.12_cpu_0
pytorch-lightning=2.4.0=pyhd8ed1ab_0
pytorch-mutex=1.0=cpu
pytz=2024.2=pypi_0
pyyaml=6.0.2=py312h827c3e9_0
qt-main=5.15.2=h19c9488_10
requests=2.32.3=py312haa95532_0
safetensors=0.4.4=py312h1429478_0
scipy=1.14.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.1.0=py312haa95532_0
six=1.16.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.2=py312haa95532_0
tbb=2021.8.0=h59b6b97_0
timm=1.0.11=pyhd8ed1ab_0
tk=8.6.14=h0416ee5_0
torchaudio=2.4.1=py312_cpu
torchmetrics=1.4.2=pyhd8ed1ab_0
torchvision=0.19.1=py312_cpu
tqdm=4.66.5=pyhd8ed1ab_0
typing-extensions=4.11.0=py312haa95532_0
typing_extensions=4.11.0=py312haa95532_0
tzdata=2024.2=pypi_0
ucrt=10.0.22621.0=h57928b3_1
ultralytics=8.3.10=pypi_0
ultralytics-thop=2.0.9=pypi_0
urllib3=2.2.3=py312haa95532_0
vc=14.40=h2eaa2aa_1
vc14_runtime=14.40.33810=hcc2c482_22
vs2015_runtime=14.40.33810=h3bf8584_22
werkzeug=3.0.3=py312haa95532_0
wheel=0.44.0=py312haa95532_0
win_inet_pton=1.1.0=py312haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
